report evaluation state on detail page
