+++ apps/hub/src/evals/components/EvaluationsTable.tsx
@@ -3,52 +3,13 @@
 import { format } from "date-fns";
 import React from "react";
 
-import { EvaluationState } from "@quri/hub-db";
-import { Table, TextTooltip } from "@quri/ui";
+import { Table } from "@quri/ui";
 
 import { StyledLink } from "@/components/ui/StyledLink";
 import { evalRunnerRoute, evaluationRoute, speclistRoute } from "@/lib/routes";
 
 import { type EvaluationSummaryDTO } from "../data/summaryEvals";
-
-// Helper function to render the state with appropriate styling
-function renderState(
-  state: EvaluationState,
-  errorMsg?: string | null
-): React.ReactNode {
-  let className = "";
-  let label = state;
-
-  switch (state) {
-    case "Pending":
-      className = "bg-yellow-100 text-yellow-800";
-      break;
-    case "Running":
-      className = "bg-blue-100 text-blue-800";
-      break;
-    case "Completed":
-      className = "bg-green-100 text-green-800";
-      break;
-    case "Failed":
-      className = "bg-red-100 text-red-800";
-      break;
-  }
-
-  const stateElement = (
-    <span
-      className={`inline-flex rounded-full px-2 text-xs font-semibold leading-5 ${className}`}
-    >
-      {label}
-    </span>
-  );
-
-  // If there's an error message and state is Failed, show it in a tooltip
-  if (state === "Failed" && errorMsg) {
-    return <TextTooltip text={errorMsg}>{stateElement}</TextTooltip>;
-  }
-
-  return stateElement;
-}
+import { EvaluationStateDisplay } from "./EvaluationStateDisplay";
 
 interface EvaluationsTableProps {
   evaluations: EvaluationSummaryDTO[];
@@ -105,7 +66,10 @@ export function EvaluationsTable({
               {format(new Date(evaluation.createdAt), "MMM d, yyyy h:mm a")}
             </Table.Cell>
             <Table.Cell>
-              {renderState(evaluation.state, evaluation.errorMsg)}
+              <EvaluationStateDisplay
+                state={evaluation.state}
+                errorMsg={evaluation.errorMsg}
+              />
             </Table.Cell>
             <Table.Cell>
               <StyledLink
